projet/requirements.txt
replicate
pillow
requests
python-dotenv
pyyaml